+++ src/Attributes/Numerus/Ceil.php
@@ -7,7 +7,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Cline\Struct\Attributes;
+namespace Cline\Struct\Attributes\Numerus;
 
 use Attribute;
 use Cline\Struct\Casts\NumericCast;
@@ -16,13 +16,13 @@ use Cline\Struct\Contracts\ProvidesCastClassInterface;
 use RoundingMode;
 
 /**
- * Rounds a numeric property toward positive infinity.
+ * Rounds a numeric property up to the nearest integer.
  *
  * @author Brian Faust <brian@cline.sh>
  * @psalm-immutable
  */
 #[Attribute(Attribute::TARGET_PROPERTY)]
-final readonly class RoundCeiling implements ConfiguresNumericRoundingInterface, ProvidesCastClassInterface
+final readonly class Ceil implements ConfiguresNumericRoundingInterface, ProvidesCastClassInterface
 {
     public function __construct(
         public int $precision = 0,